requirements.txt
alembic==1.5.5
astroid==2.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3